refactor: Move requestModifier into module
